+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ForegroundNotificationManager.kt
@@ -33,6 +33,7 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
 
     private val badgeColor = service.resources.getColor(R.color.colorPrimary)
 
+    private var onForeground = false
     private var reconnecting = false
     private var showingReconnecting = false
 
@@ -47,6 +48,9 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
 
             updateNotification()
         }
+
+    private val shouldBeOnForeground
+        get() = lockedToForeground || !(tunnelState is TunnelState.Disconnected)
 
     private val notificationText: Int
         get() {
@@ -138,10 +142,17 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
 
     var onConnect: (() -> Unit)? = null
     var onDisconnect: (() -> Unit)? = null
+
     var loggedIn = false
         set(value) {
             field = value
             updateNotification()
+        }
+
+    var lockedToForeground = false
+        set(value) {
+            field = value
+            updateNotificationForegroundStatus()
         }
 
     init {
@@ -155,9 +166,9 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
 
             registerReceiver(connectReceiver, connectFilter, PERMISSION_TUNNEL_ACTION, null)
             registerReceiver(disconnectReceiver, disconnectFilter, PERMISSION_TUNNEL_ACTION, null)
-
-            startForeground(FOREGROUND_NOTIFICATION_ID, buildNotification())
         }
+
+        updateNotification()
     }
 
     fun onDestroy() {
@@ -166,8 +177,6 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
         service.apply {
             unregisterReceiver(connectReceiver)
             unregisterReceiver(disconnectReceiver)
-
-            stopForeground(true)
         }
 
         notificationManager.cancel(FOREGROUND_NOTIFICATION_ID)
@@ -187,6 +196,20 @@ class ForegroundNotificationManager(val service: Service, val connectionProxy: C
     private fun updateNotification() {
         if (!reconnecting || !showingReconnecting) {
             notificationManager.notify(FOREGROUND_NOTIFICATION_ID, buildNotification())
+        }
+
+        updateNotificationForegroundStatus()
+    }
+
+    private fun updateNotificationForegroundStatus() {
+        if (shouldBeOnForeground != onForeground) {
+            if (shouldBeOnForeground) {
+                service.startForeground(FOREGROUND_NOTIFICATION_ID, buildNotification())
+                onForeground = true
+            } else if (!shouldBeOnForeground) {
+                service.stopForeground(Service.STOP_FOREGROUND_DETACH)
+                onForeground = false
+            }
         }
     }